# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/scholars/{scholarSlug}/series": {
      "get": {
        "operationId": "SeriesController_list",
        "parameters": [
          {
            "name": "scholarSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesViewDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "List published series for a scholar",
        "tags": [
          "Series"
        ]
      }
    },
    "/scholars/{scholarSlug}/series/{slug}": {
      "get": {
        "operationId": "SeriesController_get",
        "parameters": [
          {
            "name": "scholarSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesViewDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a published series by slug",
        "tags": [
          "Series"
        ]
      }
    },
    "/scholars/{scholarSlug}/series/upsert": {
      "post": {
        "operationId": "SeriesController_upsert",
        "parameters": [
          {
            "name": "scholarSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSeriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesViewDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Upsert a series by slug for a scholar",
        "tags": [
          "Series"
        ]
      }
    },
...
    "/series/{id}": {
      "get": {
        "operationId": "SeriesPublicController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesViewDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a published series by id",
        "tags": [
          "Series"
        ]
      }
    },
...
  },
  "info": {
    "title": "Salafi Durus API",
    "description": "Backend API for Salafi Durus",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "requestId": {
            "type": "string",
            "example": "c0a8012e-3d3b-4e72-9e5f-2d7b0f9d3d4a"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-01-28T12:34:56.789Z"
          },
          "path": {
            "type": "string",
            "example": "/health"
          }
        },
        "required": [
          "statusCode",
          "message",
          "requestId",
          "timestamp",
          "path"
        ]
      },
...
      "SeriesViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "scholarId": {
            "type": "string"
          },
          "collectionId": {
            "type": "string",
            "description": "Optional parent collection id"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "review",
              "published",
              "archived"
            ]
          },
          "orderIndex": {
            "type": "number"
          },
          "deletedAt": {
            "type": "string"
          },
          "deleteAfterAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "scholarId",
          "slug",
          "title",
          "status",
          "createdAt"
        ]
      },
      "UpsertSeriesDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Stable, URL-safe slug unique per scholar"
          },
          "title": {
            "type": "string",
            "description": "Series title"
          },
          "description": {
            "type": "string",
            "description": "Series description"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "review",
              "published",
              "archived"
            ],
            "default": "draft"
          },
          "orderIndex": {
            "type": "number",
            "description": "Optional ordering within scholar/collection"
          },
          "collectionSlug": {
            "type": "string",
            "description": "Optional parent collection slug (belongs to same scholar)",
            "example": "selected-fatawa"
          },
          "deletedAt": {
            "type": "string",
            "description": "Soft delete time ISO string"
          },
          "deleteAfterAt": {
            "type": "string",
            "description": "Auto delete time ISO string"
          }
        },
        "required": [
          "slug",
          "title"
        ]
      },
...
    }
  }
}